Reorganize memory_state to avoid giant test module
